// admin/src/api/tag/content-types/tag/tag.ts
// Interface automatically generated by schemas-to-ts

import { Blog } from '../../../blog/content-types/blog/blog';
import { Researches } from '../../../researches/content-types/researches/researches';
import { Blog_Plain } from '../../../blog/content-types/blog/blog';
import { Researches_Plain } from '../../../researches/content-types/researches/researches';
import { AdminPanelRelationPropertyModification } from '../../../../common/schemas-to-ts/AdminPanelRelationPropertyModification';

export interface Tag {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    name?: string;
    slug?: string;
    blogs?: { data: Blog[] };
    research?: { data: Researches[] };
  };
}
export interface Tag_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  name?: string;
  slug?: string;
  blogs?: Blog_Plain[];
  research?: Researches_Plain[];
}

export interface Tag_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  name?: string;
  slug?: string;
  blogs?: number[];
  research?: number[];
}

export interface Tag_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  name?: string;
  slug?: string;
  blogs?: AdminPanelRelationPropertyModification<Blog_Plain>;
  research?: AdminPanelRelationPropertyModification<Researches_Plain>;
}