// admin/src/api/category/content-types/category/category.ts
// Interface automatically generated by schemas-to-ts

import { Blog } from '../../../blog/content-types/blog/blog'
import { Blog_Plain } from '../../../blog/content-types/blog/blog'
import { AdminPanelRelationPropertyModification } from '../../../../common/schemas-to-ts/AdminPanelRelationPropertyModification'

export interface Category {
  id: number
  attributes: {
    createdAt: Date
    updatedAt: Date
    publishedAt?: Date
    name?: string
    slug?: string
    blogs: { data: Blog[] }
  }
}
export interface Category_Plain {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  name?: string
  slug?: string
  blogs: Blog_Plain[]
}

export interface Category_NoRelations {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  name?: string
  slug?: string
  blogs: number[]
}

export interface Category_AdminPanelLifeCycle {
  id: number
  createdAt: Date
  updatedAt: Date
  publishedAt?: Date
  name?: string
  slug?: string
  blogs: AdminPanelRelationPropertyModification<Blog_Plain>
}